source/backend/requirements.txt
# Web framework and API dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
starlette>=0.14.2

# AWS dependencies
boto3==1.38.9
botocore==1.38.9

# Data validation and serialization
pydantic>=1.8.0

# Environment and configuration
python-dotenv>=0.19.0

# Utility packages
python-dateutil>=2.8.2
requests>=2.28.0

# Logging and monitoring
structlog>=21.1.0

# HTTP middleware
aiohttp>=3.8.0

# Authentication and security
PyJWT>=2.8.0
cryptography>=41.0.0

# Audio dependencies
pyaudio>=0.2.13
rx>=3.2.0
smithy-aws-core>=0.0.1
pytz
aws_sdk_bedrock_runtime==0.0.2

websockets==15.0.1
Pillow>=10.0.0  # Python Imaging Library for image processing
python-multipart

psutil